+++ test/observer_test.dart
@@ -42,7 +42,7 @@ void main() {
             home: Material(
               child: MultiAtomObserver(
                 atoms: [index, character],
-                builder: (context, state) {
+                builder: (context) {
                   final name = character.value;
                   final id = index.value;
                   return Text("$id: $name");