+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+- Fix: Removed unexisting state parameter reference from MultiAtomObserver
+- Added removeListeners() to AtomNotifier, which is meant to remove all listeners
+
 ## 1.0.0
 - Added default reactive components
 - Added default application states